Added yes, no and average button to habit component. Habit component also shows times completed, days and score
