+++ life-gpa-frontend/src/Dashboard/dashboard.js
@@ -11,6 +11,7 @@ import OverallGPA from './Components/OverallGPA'
 import HabitCard from './Components/habitCard'
 import Buttons from './Components/buttonComponent'
 
+
 const Dashboard = () => {
     const [habitTracker, setHabitTracker ] = useState( [] )
     const addToHabitTracker = habit => {
